Allow use without a GitHub token.
Do not mandate the use of a GitHub token to generate the release pages.
Try to find a token in a file called '.token' in the top-level
directory, but proceed without token if such a file does not exist. The
script should not need more than one API call for every 100 releases, so
it will not hit the rate limit for unauthenticated users until we reach
the 6000th release.